move to utils/volume
